Versao 1.0.12

+++ OpenSigFiscal/src/br/com/opensig/fiscal/server/sped/bloco0/Dados0220.java
@@ -5,7 +5,7 @@ import br.com.opensig.fiscal.server.sped.Bean;
 public class Dados0220 extends Bean {
 
 	private String unid_conv;
-	private int fat_conv;
+	private Double fat_conv;
 
 	public Dados0220() {
 		reg = "0220";
@@ -19,11 +19,11 @@ public class Dados0220 extends Bean {
 		this.unid_conv = unid_conv;
 	}
 
-	public int getFat_conv() {
+	public Double getFat_conv() {
 		return fat_conv;
 	}
 
-	public void setFat_conv(int fat_conv) {
+	public void setFat_conv(Double fat_conv) {
 		this.fat_conv = fat_conv;
 	}
 }
